requirements.txt
flask>=3.0
Pillow>=10.0
numpy>=1.24
opencv-python-headless>=4.8
requests>=2.31
gunicorn>=21.2.0
